backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.50.0
httpx>=0.27.0,<0.28.0
youtube-transcript-api>=1.0.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
google-api-python-client>=2.100.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0
